+++ crates/zng-wgt-scroll/src/node.rs
@@ -561,7 +561,11 @@ pub fn zoom_commands_node(child: impl IntoUiNode) -> UiNode {
         let scroll = WIDGET.info().scroll_info().unwrap();
         let viewport = (scroll.viewport_size() + scroll.joiner_size()).to_f32(); // viewport without scrollbars
         let content = scroll.content().size.max(PxSize::splat(Px(1))).to_f32() / scroll.zoom_scale();
-        (viewport.width / content.width).min(viewport.height / content.height).fct()
+        let scale = (viewport.width / content.width).min(viewport.height / content.height).fct();
+        match ZOOM_TO_FIT_MODE_VAR.get() {
+            ZoomToFitMode::Contain => scale,
+            ZoomToFitMode::ScaleDown => scale.min(1.fct()),
+        }
     }
 
     match_node(child, move |child, op| match op {
@@ -602,12 +606,9 @@ pub fn zoom_commands_node(child: impl IntoUiNode) -> UiNode {
                 });
             } else if let Some(args) = ZOOM_TO_FIT_CMD.scoped(scope).on(update) {
                 args.handle_enabled(&zoom_to_fit, |args| {
-                    let mut scale = fit_scale();
+                    let scale = fit_scale();
                     if let Some(p) = ZoomToFitRequest::from_args(args) {
-                        if p.scale_down {
-                            scale = scale.min(1.fct());
-                        }
-                        if p.skip_animation {
+                        if dbg!(p.skip_animation) {
                             SMOOTH_SCROLLING_VAR.with_context_var(zng_var::ContextInitHandle::current(), false, || {
                                 SCROLL.chase_zoom(|_| scale);
                             });