+++ crates/zng/src/scroll.rs
@@ -63,10 +63,10 @@
 
 pub use zng_wgt_scroll::{
     LazyMode, SCROLL, Scroll, ScrollBarArgs, ScrollFrom, ScrollInfo, ScrollMode, ScrollUnitsMix, Scrollbar, ScrollbarFnMix,
-    SmoothScrolling, Thumb, WidgetInfoExt, alt_factor, auto_hide_extra, clip_to_viewport, define_viewport_unit, h_line_unit, h_page_unit,
-    h_scrollbar_fn, h_wheel_unit, lazy, line_units, max_zoom, min_zoom, mode, mouse_pan, overscroll_color, page_units,
+    SmoothScrolling, Thumb, WidgetInfoExt, ZoomToFitMode, alt_factor, auto_hide_extra, clip_to_viewport, define_viewport_unit, h_line_unit,
+    h_page_unit, h_scrollbar_fn, h_wheel_unit, lazy, line_units, max_zoom, min_zoom, mode, mouse_pan, overscroll_color, page_units,
     scroll_to_focused_mode, scrollbar_fn, scrollbar_joiner_fn, smooth_scrolling, v_line_unit, v_page_unit, v_scrollbar_fn, v_wheel_unit,
-    wheel_units, zoom_origin, zoom_size_only, zoom_touch_origin, zoom_wheel_origin, zoom_wheel_unit,
+    wheel_units, zoom_origin, zoom_size_only, zoom_to_fit_mode, zoom_touch_origin, zoom_wheel_origin, zoom_wheel_unit,
 };
 
 /// Scrollbar thumb widget.